Move to Mixin and fix sizing of Project card
